# docs/swagger.yaml
basePath: /
definitions:
...
  appsettings.settingsExport:
    properties:
      exported_at:
        type: string
      settings:
        additionalProperties:
          type: string
        type: object
      version:
        type: integer
    type: object
  appsettings.settingsPreviewResponse:
    properties:
      rows:
        items:
          $ref: '#/definitions/appsettings.settingsPreviewRow'
        type: array
      unknown:
        items:
          type: string
        type: array
    type: object
  appsettings.settingsPreviewRow:
    properties:
      current:
        type: string
      incoming:
        type: string
      key:
        type: string
    type: object
  audit.AuditLogListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/audit.AuditLogResponse'
        type: array
      total:
        type: integer
    type: object
  audit.AuditLogResponse:
    properties:
      actor_id:
        type: string
      actor_username:
        type: string
      created_at:
        type: string
      detail:
        type: string
      event:
        type: string
      id:
        type: string
      ip_address:
        type: string
      target_id:
        type: string
      target_type:
        type: string
    type: object
...
  model.ApiError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
host: localhost:9999
info:
  contact: {}
  description: Authentication Service
  title: Autentico OIDC
  version: "1.0"
paths:
...
  /admin/api/audit-logs:
    get:
      description: Returns a paginated list of audit events with optional filters.
      parameters:
      - description: Filter by event type
        in: query
        name: event
        type: string
      - description: Filter by actor user ID
        in: query
        name: actor_id
        type: string
      - description: Page size (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audit.AuditLogListResponse'
      security:
      - BearerAuth: []
      summary: List audit log events
      tags:
      - admin
...
  /admin/api/settings:
    get:
      description: Retrieve all system settings (except sensitive values).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get system settings
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update multiple settings by key-value pairs.
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - BearerAuth: []
      summary: Update system settings
      tags:
      - admin
  /admin/api/settings/export:
    get:
      description: Export all settings as a JSON file (smtp_password excluded).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/appsettings.settingsExport'
      security:
      - BearerAuth: []
      summary: Export settings
      tags:
      - admin
  /admin/api/settings/import/apply:
    post:
      consumes:
      - application/json
      description: Applies imported settings, skipping unknown keys and protected
        fields.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Apply settings import
      tags:
      - admin
  /admin/api/settings/import/preview:
    post:
      consumes:
      - application/json
      description: Returns a diff of current vs incoming values for all known keys,
        plus unknown keys.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/appsettings.settingsPreviewResponse'
      security:
      - BearerAuth: []
      summary: Preview settings import
      tags:
      - admin
  /admin/api/settings/test-smtp:
    post:
      description: Sends a test email to the currently authenticated admin's registered
        email address.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - BearerAuth: []
      summary: Test SMTP configuration
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"